+++ crates/ruma-macros/src/identifiers/id_dst.rs
@@ -97,6 +97,13 @@ impl IdDst {
             }
 
             #[automatically_derived]
+            impl #impl_generics ::std::borrow::Borrow<#str> for #id {
+                fn borrow(&self) -> &#str {
+                    self.as_str()
+                }
+            }
+
+            #[automatically_derived]
             impl #impl_generics ::std::convert::AsRef<#id> for #id {
                 fn as_ref(&self) -> &#id {
                     self
@@ -359,6 +366,13 @@ impl IdDst {
             impl #impl_generics ::std::borrow::Borrow<#id> for #owned_id {
                 fn borrow(&self) -> &#id {
                     self.as_ref()
+                }
+            }
+
+            #[automatically_derived]
+            impl #impl_generics ::std::borrow::Borrow<#str> for #owned_id {
+                fn borrow(&self) -> &#str {
+                    self.as_inner_str()
                 }
             }