+++ crates/ruma-common/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 Bug fixes:
+
 - Properly check the room type filter in `ruma_common::directory::Filter`'s `is_empty` function.
 
 Breaking changes: